refactor(form): move context creation to index & remove tokens
