--- packages/form/src/tokens/lifecycle.js
@@ -1,2 +1,0 @@
-export const input = Symbol('input');
-export const meta = Symbol('meta');
